+++ src/styleguide/demo-pages/elements/r-button.vue
@@ -53,7 +53,7 @@
 </template>
 
 <script lang="ts">
-  import lVasLayout from '@valantic/vue-styleguide/src/layouts/l-vas-layout.vue';
+  import { lVasLayout } from '@valantic/vue-styleguide/layouts';
   import { defineComponent } from 'vue';
   import eButton, { BUTTON_STYLE_VARIANTS, BUTTON_WIDTH_VARIANTS } from '@/elements/e-button.vue';