+++ src/styleguide/demo-pages/elements/r-loading-indicator.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script lang="ts">
-  import lVasLayout from '@valantic/vue-styleguide/src/layouts/l-vas-layout.vue';
+  import { lVasLayout } from '@valantic/vue-styleguide/layouts';
   import { defineComponent } from 'vue';
   import eLoadingIndicator from '@/elements/e-loading-indicator.vue';