+++ src/styleguide/demo-pages/elements/r-wysiwyg.vue
@@ -112,7 +112,7 @@
 </template>
 
 <script lang="ts">
-  import lVasLayout from '@valantic/vue-styleguide/src/layouts/l-vas-layout.vue';
+  import { lVasLayout } from '@valantic/vue-styleguide/layouts';
   import { defineComponent } from 'vue';
 
   export default defineComponent({