+++ src/styleguide/demo-pages/general/r-colors.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script lang="ts">
-  import lVasLayout from '@valantic/vue-styleguide/src/layouts/l-vas-layout.vue';
+  import { lVasLayout } from '@valantic/vue-styleguide/layouts';
 
   export default {
     name: 'r-colors',