+++ src/styleguide/demo-pages/general/r-icons.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-  import lVasLayout from '@valantic/vue-styleguide/src/layouts/l-vas-layout.vue';
+  import { lVasLayout } from '@valantic/vue-styleguide/layouts';
   import { defineComponent } from 'vue';
   import sIconFinder from '@/styleguide/components/s-icon-finder.vue';